Import of XFree86 4.3.0.

+++ xfree/xc/lib/Xtst/XRecord.c
@@ -50,17 +50,17 @@ from The Open Group.
 /*
  * By Stephen Gildea, X Consortium, and Martha Zimet, NCD.
  */
-/* $XFree86: xc/lib/Xtst/XRecord.c,v 1.4 2001/12/14 19:56:39 dawes Exp $ */
+/* $XFree86: xc/lib/Xtst/XRecord.c,v 1.6 2002/10/16 00:37:33 dawes Exp $ */
 
 #include <stdio.h>
 #include <assert.h>
 #define NEED_EVENTS
 #define NEED_REPLIES
 #include <X11/Xlibint.h>
-#include "Xext.h"
+#include <X11/extensions/Xext.h>
 #include <X11/Xtrans.h>
-#include "extutil.h"
-#include "recordstr.h"
+#include <X11/extensions/extutil.h>
+#include <X11/extensions/recordstr.h>
 
 extern unsigned long _XSetLastRequestRead();
 
@@ -822,7 +822,7 @@ parse_reply_call_callback(dpy, info, rep, reply, callback, closure)
 	    if (current_index + datum_bytes > rep->length << 2)
 		fprintf(stderr,
 			"XRecord: %lu-byte reply claims %d-byte element (seq %lu)\n",
-			rep->length << 2, current_index + datum_bytes,
+			(long)rep->length << 2, current_index + datum_bytes,
 			dpy->last_request_read);
 	    /*
 	     * This assignment (and indeed the whole buffer sharing
